@@ -18,24 +18,24 @@
     }
   ],
   "require": {
-    "php": "^7.1 || ^8.0",
-    "symfony/http-kernel": "^4.4|^5.0",
+    "php": ">=8.0",
+    "symfony/http-kernel": "^5.0|^6.0",
     "ext-json": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "^5.0",
-    "mockery/mockery": "^1.0",
-    "symfony/validator": "^4.4|^5.0",
-    "symfony/form": "^4.4|^5.0"
+    "phpunit/phpunit": "^9.5",
+    "mockery/mockery": "^1.4",
+    "symfony/validator": "^5.0|^6.0",
+    "symfony/form": "^5.0|^6.0"
   },
   "autoload": {
     "psr-4": {
-      "MediaMonks\\RestApi\\": "src/"
+      "MediaMonks\\RestApi\\": "src"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "tests\\MediaMonks\\RestApi\\": "tests/src/"
+      "MediaMonks\\RestApi\\Tests\\": "tests"
     }
   },
   "extra": {
