@@ -4,10 +4,7 @@ namespace MediaMonks\RestApi\Request;
 
 class PathRequestMatcher extends RegexRequestMatcher
 {
-    /**
-     * @param string $path
-     */
-    public function __construct($path)
+    public function __construct(string $path)
     {
         parent::__construct([sprintf('~^%s~', $path), sprintf('~^%s/~', $path)]);
     }
